requirements.txt
matplotlib==3.1.2
numpy==1.18.0
pyyaml==19.12.0
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.3
